test: set up testing for layered cache
