+++ rust/property-defs-rs/src/cache/layered_cache.rs
@@ -2,17 +2,18 @@ use std::sync::Arc;
 use quick_cache::sync::Cache as InMemoryCache;
 use crate::types::Update;
 use crate::errors::CacheError;
-use super::{SecondaryCache, CacheOperations};
+use crate::cache::secondary_cache::SecondaryCacheOperations;
+use crate::cache::secondary_cache::SecondaryCache;
 use tracing::warn;
 
 #[derive(Clone)]
-pub struct LayeredCache {
+pub struct LayeredCache<S: SecondaryCacheOperations = SecondaryCache> {
     memory: Arc<InMemoryCache<Update, ()>>,
-    secondary: SecondaryCache,
+    secondary: S,
 }
 
-impl LayeredCache {
-    pub fn new(memory: Arc<InMemoryCache<Update, ()>>, secondary: SecondaryCache) -> Self {
+impl<S: SecondaryCacheOperations> LayeredCache<S> {
+    pub fn new(memory: Arc<InMemoryCache<Update, ()>>, secondary: S) -> Self {
         Self { memory, secondary }
     }
 
@@ -66,63 +67,5 @@ impl LayeredCache {
 
     pub fn remove(&self, key: &Update) -> Option<()> {
         self.memory.remove(key).map(|_| ())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::cache::noop_cache::NoOpCache;
-    use crate::types::{Update, EventDefinition};
-    use chrono::Utc;
-
-    #[tokio::test]
-    async fn test_layered_cache_basic() {
-        let memory = Arc::new(InMemoryCache::new(1000));
-        let secondary = SecondaryCache::NoOp(NoOpCache::new());
-        let cache = LayeredCache::new(memory, secondary);
-
-        let events: Vec<Update> = (0..3)
-            .map(|i| {
-                Update::Event(EventDefinition {
-                    name: format!("test_{}", i),
-                    team_id: 1,
-                    project_id: 1,
-                    last_seen_at: Utc::now(),
-                })
-            })
-            .collect();
-
-        // First insert the events
-        cache.insert_batch(events.clone()).await;
-
-        // Then verify none of them are returned by filter_cached_updates
-        let not_in_cache = cache.filter_cached_updates(events.clone()).await;
-        assert_eq!(not_in_cache.len(), 0);
-    }
-
-    #[tokio::test]
-    async fn test_layered_cache_large() {
-        let memory = Arc::new(InMemoryCache::new(10000));
-        let secondary = SecondaryCache::NoOp(NoOpCache::new());
-        let cache = LayeredCache::new(memory, secondary);
-
-        let events: Vec<Update> = (0..3)
-            .map(|i| {
-                Update::Event(EventDefinition {
-                    name: format!("test_{}", i),
-                    team_id: 1,
-                    project_id: 1,
-                    last_seen_at: Utc::now(),
-                })
-            })
-            .collect();
-
-        // First insert the events
-        cache.insert_batch(events.clone()).await;
-
-        // Then verify none of them are returned by filter_cached_updates
-        let not_in_cache = cache.filter_cached_updates(events.clone()).await;
-        assert_eq!(not_in_cache.len(), 0);
     }
 }
