+++ rust/property-defs-rs/tests/mod.rs
@@ -1,0 +1,5 @@
+mod cache;
+mod queries;
+mod types;
+mod updates;
+mod v2_batch_ingestion;
